Refactor ballot upload button into component
