+++ packages/admin-app/src/lang/en.json
@@ -230,6 +230,7 @@
     "adduser.title": "Add User",
     "adduser.error": "An error occurred while adding the user",
 
+    "uploadballots.selectfiles": "Select a Ballot json file",
     "uploadballots.title": "Upload Ballots",
     "uploadballots_error_noballots": "The ballots json file looks wrong, it should contain at least one ballot",
     "uploadballots_error_ballotsinvalid": "At least one of the ballots in the ballots json file looks incorrect, ensure they all have an object_id, state, and contests"
